Take ImageViewer out of amp-image-lightbox so it can be reused
