+++ src/image-viewer.js
@@ -1,0 +1,637 @@
+/**
+ * This class is responsible providing all operations necessary for viewing
+ * an image, such as full-bleed display, zoom and pan, etc.
+ * @package  Visible for testing only!
+ * TODO(dvoytenko): move to the separate file once build system is ready.
+ */
+import {Animation} from './animation';
+import {Gestures} from './gesture';
+import {
+  DoubletapRecognizer,
+  SwipeXYRecognizer,
+  TapRecognizer,
+  TapzoomRecognizer,
+} from './gesture-recognizers';
+import {bezierCurve} from './curve';
+import {isLoaded} from './event-helper';
+import {Services} from './services';
+import {continueMotion} from './motion';
+import {srcsetFromElement} from './srcset';
+import {
+  layoutRectFromDomRect,
+  layoutRectLtwh,
+  moveLayoutRect,
+} from './layout-rect';
+import * as st from './style';
+import * as tr from './transition';
+
+/** @private @const {!../../../src/curve.CurveDef} */
+const PAN_ZOOM_CURVE_ = bezierCurve(0.4, 0, 0.2, 1.4);
+
+export class ImageViewer {
+  /**
+   * @param {!AmpImageLightbox} lightbox
+   * @param {!Window} win
+   * @param {!function(T, number=):Promise<T>} loadPromise
+   * @template T
+   */
+  constructor(lightbox, win, loadPromise) {
+    /** @private {!AmpImageLightbox} */
+    this.lightbox_ = lightbox;
+
+    /** @const {!Window} */
+    this.win = win;
+
+    /** @private {function(T, number=):Promise<T>} */
+    this.loadPromise_ = loadPromise;
+
+    /** @private {!Element} */
+    this.viewer_ = lightbox.element.ownerDocument.createElement('div');
+    this.viewer_.classList.add('i-amphtml-image-lightbox-viewer');
+
+    /** @private {!Element} */
+    this.image_ = lightbox.element.ownerDocument.createElement('img');
+    this.image_.classList.add('i-amphtml-image-lightbox-viewer-image');
+    this.viewer_.appendChild(this.image_);
+
+    /** @private {./srcset.Srcset} */
+    this.srcset_ = null;
+
+    /** @private {!Object} */
+    this.ariaAttributes_ = {
+      'alt': null,
+      'aria-label': null,
+      'aria-labelledby': null,
+    };
+
+    /** @private {number} */
+    this.sourceWidth_ = 0;
+
+    /** @private {number} */
+    this.sourceHeight_ = 0;
+
+    /** @private {./src/layout-rect.LayoutRectDef} */
+    this.viewerBox_ = layoutRectLtwh(0, 0, 0, 0);
+
+    /** @private {./src/layout-rect.LayoutRectDef} */
+    this.imageBox_ = layoutRectLtwh(0, 0, 0, 0);
+
+    /** @private {number} */
+    this.scale_ = 1;
+    /** @private {number} */
+    this.startScale_ = 1;
+    /** @private {number} */
+    this.maxSeenScale_ = 1;
+    /** @private {number} */
+    this.minScale_ = 1;
+    /** @private {number} */
+    this.maxScale_ = 2;
+
+    /** @private {number} */
+    this.startX_ = 0;
+    /** @private {number} */
+    this.startY_ = 0;
+    /** @private {number} */
+    this.posX_ = 0;
+    /** @private {number} */
+    this.posY_ = 0;
+    /** @private {number} */
+    this.minX_ = 0;
+    /** @private {number} */
+    this.minY_ = 0;
+    /** @private {number} */
+    this.maxX_ = 0;
+    /** @private {number} */
+    this.maxY_ = 0;
+
+    /** @private {?../../../src/motion.Motion} */
+    this.motion_ = null;
+
+    this.setupGestures_();
+  }
+
+  /**
+   * Returns the root element of the image viewer.
+   * @return {!Element}
+   */
+  getElement() {
+    return this.viewer_;
+  }
+
+  /**
+   * Returns the img element of the image viewer.
+   * @return {!Element}
+   */
+  getImage() {
+    return this.image_;
+  }
+
+  /**
+   * Returns the boundaries of the viewer.
+   * @return {!../../../src/layout-rect.LayoutRectDef}
+   */
+  getViewerBox() {
+    return this.viewerBox_;
+  }
+
+  /**
+   * Returns the boundaries of the image element.
+   * @return {!../../../src/layout-rect.LayoutRectDef}
+   */
+  getImageBox() {
+    return this.imageBox_;
+  }
+
+  /**
+   * Returns the boundaries of the image element with the offset if it was
+   * moved by a gesture.
+   * @return {!../../../src/layout-rect.LayoutRectDef}
+   */
+  getImageBoxWithOffset() {
+    if (this.posX_ == 0 && this.posY_ == 0) {
+      return this.imageBox_;
+    }
+    return moveLayoutRect(this.imageBox_, this.posX_, this.posY_);
+  }
+
+  /**
+   * Resets the image viewer to the initial state.
+   */
+  reset() {
+    this.image_.setAttribute('src', '');
+    Object.keys(this.ariaAttributes_).forEach(key => {
+      this.image_.removeAttribute(key);
+      this.ariaAttributes_[key] = null;
+    });
+    this.image_.removeAttribute('aria-describedby');
+    this.srcset_ = null;
+    this.imageBox_ = layoutRectLtwh(0, 0, 0, 0);
+    this.sourceWidth_ = 0;
+    this.sourceHeight_ = 0;
+
+    this.maxSeenScale_ = 1;
+    this.scale_ = 1;
+    this.startScale_ = 1;
+    this.maxScale_ = 2;
+
+    this.startX_ = 0;
+    this.startY_ = 0;
+    this.posX_ = 0;
+    this.posY_ = 0;
+    this.minX_ = 0;
+    this.minY_ = 0;
+    this.maxX_ = 0;
+    this.maxY_ = 0;
+
+    if (this.motion_) {
+      this.motion_.halt();
+    }
+    this.motion_ = null;
+  }
+
+  /**
+   * Initializes the image viewer to the target image element such as
+   * "amp-img". The target image element may or may not yet have the img
+   * element initialized.
+   * @param {!Element} sourceElement
+   * @param {?Element} sourceImage
+   */
+  init(sourceElement, sourceImage) {
+    this.sourceWidth_ = sourceElement./*OK*/offsetWidth;
+    this.sourceHeight_ = sourceElement./*OK*/offsetHeight;
+    this.srcset_ = srcsetFromElement(sourceElement);
+
+    Object.keys(this.ariaAttributes_).forEach(key => {
+      this.ariaAttributes_[key] = sourceElement.getAttribute(key);
+      if (this.ariaAttributes_[key]) {
+        this.image_.setAttribute(key, this.ariaAttributes_[key]);
+      }
+    });
+
+    if (sourceImage && isLoaded(sourceImage) && sourceImage.src) {
+      // Set src provisionally to the known loaded value for fast display.
+      // It will be updated later.
+      this.image_.setAttribute('src', sourceImage.src);
+    }
+  }
+
+  /**
+   * Measures the image viewer and image sizes and positioning.
+   * @return {!Promise}
+   */
+  measure() {
+    this.viewerBox_ = layoutRectFromDomRect(this.viewer_
+        ./*OK*/getBoundingClientRect());
+
+    const sf = Math.min(this.viewerBox_.width / this.sourceWidth_,
+        this.viewerBox_.height / this.sourceHeight_);
+    let width = Math.min(this.sourceWidth_ * sf, this.viewerBox_.width);
+    let height = Math.min(this.sourceHeight_ * sf, this.viewerBox_.height);
+
+    // TODO(dvoytenko): This is to reduce very small expansions that often
+    // look like a stutter. To be evaluated if this is still the right
+    // idea.
+    if (width - this.sourceWidth_ <= 16) {
+      width = this.sourceWidth_;
+      height = this.sourceHeight_;
+    }
+
+    this.imageBox_ = layoutRectLtwh(
+        Math.round((this.viewerBox_.width - width) / 2),
+        Math.round((this.viewerBox_.height - height) / 2),
+        Math.round(width),
+        Math.round(height));
+
+    st.setStyles(this.image_, {
+      top: st.px(this.imageBox_.top),
+      left: st.px(this.imageBox_.left),
+      width: st.px(this.imageBox_.width),
+      height: st.px(this.imageBox_.height),
+    });
+
+    // Reset zoom and pan.
+    this.startScale_ = this.scale_ = 1;
+    this.startX_ = this.posX_ = 0;
+    this.startY_ = this.posY_ = 0;
+    this.updatePanZoomBounds_(this.scale_);
+    this.updatePanZoom_();
+
+    return this.updateSrc_();
+  }
+
+  /**
+   * @return {!Promise}
+   * @private
+   */
+  updateSrc_() {
+    if (!this.srcset_) {
+      // Do not update source if the lightbox has already exited.
+      return Promise.resolve();
+    }
+    this.maxSeenScale_ = Math.max(this.maxSeenScale_, this.scale_);
+    const width = this.imageBox_.width * this.maxSeenScale_;
+    const src = this.srcset_.select(width, this.lightbox_.getDpr()).url;
+    if (src == this.image_.getAttribute('src')) {
+      return Promise.resolve();
+    }
+    // Notice that we will wait until the next event cycle to set the "src".
+    // This ensures that the already available image will show immediately
+    // and then naturally upgrade to a higher quality image.
+    return Services.timerFor(this.win).promise(1).then(() => {
+      this.image_.setAttribute('src', src);
+      return this.loadPromise_(this.image_);
+    });
+  }
+
+  /** @private */
+  setupGestures_() {
+    const gestures = Gestures.get(this.image_);
+
+    // Toggle viewer mode.
+    gestures.onGesture(TapRecognizer, () => {
+      this.lightbox_.toggleViewMode();
+    });
+
+    // Movable.
+    gestures.onGesture(SwipeXYRecognizer, e => {
+      this.onMove_(e.data.deltaX, e.data.deltaY, false);
+      if (e.data.last) {
+        this.onMoveRelease_(e.data.velocityX, e.data.velocityY);
+      }
+    });
+    gestures.onPointerDown(() => {
+      if (this.motion_) {
+        this.motion_.halt();
+      }
+    });
+
+    // Zoomable.
+    gestures.onGesture(DoubletapRecognizer, e => {
+      let newScale;
+      if (this.scale_ == 1) {
+        newScale = this.maxScale_;
+      } else {
+        newScale = this.minScale_;
+      }
+      const deltaX = this.viewerBox_.width / 2 - e.data.clientX;
+      const deltaY = this.viewerBox_.height / 2 - e.data.clientY;
+      this.onZoom_(newScale, deltaX, deltaY, true).then(() => {
+        return this.onZoomRelease_(0, 0, 0, 0, 0, 0);
+      });
+    });
+    gestures.onGesture(TapzoomRecognizer, e => {
+      this.onZoomInc_(e.data.centerClientX, e.data.centerClientY,
+          e.data.deltaX, e.data.deltaY);
+      if (e.data.last) {
+        this.onZoomRelease_(e.data.centerClientX, e.data.centerClientY,
+            e.data.deltaX, e.data.deltaY, e.data.velocityY, e.data.velocityY);
+      }
+    });
+  }
+
+  /**
+   * Returns value bound to min and max values +/- extent.
+   * @param {number} v
+   * @param {number} min
+   * @param {number} max
+   * @param {number} extent
+   * @return {number}
+   * @private
+   */
+  boundValue_(v, min, max, extent) {
+    return Math.max(min - extent, Math.min(max + extent, v));
+  }
+
+  /**
+   * Returns the scale within the allowed range with possible extent.
+   * @param {number} s
+   * @param {boolean} allowExtent
+   * @return {number}
+   * @private
+   */
+  boundScale_(s, allowExtent) {
+    return this.boundValue_(s, this.minScale_, this.maxScale_,
+        allowExtent ? 0.25 : 0);
+  }
+
+  /**
+   * Returns the X position within the allowed range with possible extent.
+   * @param {number} x
+   * @param {boolean} allowExtent
+   * @return {number}
+   * @private
+   */
+  boundX_(x, allowExtent) {
+    return this.boundValue_(x, this.minX_, this.maxX_,
+        allowExtent && this.scale_ > 1 ? this.viewerBox_.width * 0.25 : 0);
+  }
+
+  /**
+   * Returns the Y position within the allowed range with possible extent.
+   * @param {number} y
+   * @param {boolean} allowExtent
+   * @return {number}
+   * @private
+   */
+  boundY_(y, allowExtent) {
+    return this.boundValue_(y, this.minY_, this.maxY_,
+        allowExtent ? this.viewerBox_.height * 0.25 : 0);
+  }
+
+  /**
+   * Updates X/Y bounds based on the provided scale value. The min/max bounds
+   * are calculated to allow full pan of the image regardless of the scale
+   * value.
+   * @param {number} scale
+   * @private
+   */
+  updatePanZoomBounds_(scale) {
+    let maxY = 0;
+    let minY = 0;
+    const dh = this.viewerBox_.height - this.imageBox_.height * scale;
+    if (dh >= 0) {
+      minY = maxY = 0;
+    } else {
+      minY = dh / 2;
+      maxY = -minY;
+    }
+
+    let maxX = 0;
+    let minX = 0;
+    const dw = this.viewerBox_.width - this.imageBox_.width * scale;
+    if (dw >= 0) {
+      minX = maxX = 0;
+    } else {
+      minX = dw / 2;
+      maxX = -minX;
+    }
+
+    this.minX_ = minX;
+    this.minY_ = minY;
+    this.maxX_ = maxX;
+    this.maxY_ = maxY;
+  }
+
+  /**
+   * Updates pan/zoom of the image based on the current values.
+   * @private
+   */
+  updatePanZoom_() {
+    st.setStyles(this.image_, {
+      transform: st.translate(this.posX_, this.posY_) +
+          ' ' + st.scale(this.scale_),
+    });
+    if (this.scale_ != 1) {
+      this.lightbox_.toggleViewMode(true);
+    }
+  }
+
+  /**
+   * Performs a one-step or an animated motion (panning).
+   * @param {number} deltaX
+   * @param {number} deltaY
+   * @param {boolean} animate
+   * @private
+   */
+  onMove_(deltaX, deltaY, animate) {
+    const newPosX = this.boundX_(this.startX_ + deltaX, true);
+    const newPosY = this.boundY_(this.startY_ + deltaY, true);
+    this.set_(this.scale_, newPosX, newPosY, animate);
+  }
+
+  /**
+   * Performs actions once the motion gesture has been complete. The motion
+   * may continue based on the final velocity.
+   * @param {number} veloX
+   * @param {number} veloY
+   * @private
+   */
+  onMoveRelease_(veloX, veloY) {
+    const deltaY = this.posY_ - this.startY_;
+    if (this.scale_ == 1 && Math.abs(deltaY) > 10) {
+      this.lightbox_.close();
+      return;
+    }
+
+    // Continue motion.
+    this.motion_ = continueMotion(this.image_,
+        this.posX_, this.posY_, veloX, veloY,
+        (x, y) => {
+          const newPosX = this.boundX_(x, true);
+          const newPosY = this.boundY_(y, true);
+          if (Math.abs(newPosX - this.posX_) < 1 &&
+                Math.abs(newPosY - this.posY_) < 1) {
+            // Hit the wall: stop motion.
+            return false;
+          }
+          this.set_(this.scale_, newPosX, newPosY, false);
+          return true;
+        });
+
+    // Snap back.
+    this.motion_.thenAlways(() => {
+      this.motion_ = null;
+      return this.release_();
+    });
+  }
+
+  /**
+   * Performs a one-step zoom action.
+   * @param {number} centerClientX
+   * @param {number} centerClientY
+   * @param {number} deltaX
+   * @param {number} deltaY
+   * @private
+   */
+  onZoomInc_(centerClientX, centerClientY, deltaX, deltaY) {
+    const dist = Math.sqrt(deltaX * deltaX + deltaY * deltaY);
+
+    const zoomSign = Math.abs(deltaY) > Math.abs(deltaX) ?
+        Math.sign(deltaY) : Math.sign(-deltaX);
+    if (zoomSign == 0) {
+      return;
+    }
+
+    const newScale = this.startScale_ * (1 + zoomSign * dist / 100);
+    const deltaCenterX = this.viewerBox_.width / 2 - centerClientX;
+    const deltaCenterY = this.viewerBox_.height / 2 - centerClientY;
+    deltaX = Math.min(deltaCenterX, deltaCenterX * (dist / 100));
+    deltaY = Math.min(deltaCenterY, deltaCenterY * (dist / 100));
+    this.onZoom_(newScale, deltaX, deltaY, false);
+  }
+
+  /**
+   * Performs a one-step or an animated zoom action.
+   * @param {number} scale
+   * @param {number} deltaX
+   * @param {number} deltaY
+   * @param {boolean} animate
+   * @return {!Promise|undefined}
+   * @private
+   */
+  onZoom_(scale, deltaX, deltaY, animate) {
+    const newScale = this.boundScale_(scale, true);
+    if (newScale == this.scale_) {
+      return;
+    }
+
+    this.updatePanZoomBounds_(newScale);
+
+    const newPosX = this.boundX_(this.startX_ + deltaX * newScale, false);
+    const newPosY = this.boundY_(this.startY_ + deltaY * newScale, false);
+    return /** @type {!Promise|undefined} */ (
+        this.set_(newScale, newPosX, newPosY, animate));
+  }
+
+  /**
+   * Performs actions after the gesture that was performing zooming has been
+   * released. The zooming may continue based on the final velocity.
+   * @param {number} centerClientX
+   * @param {number} centerClientY
+   * @param {number} deltaX
+   * @param {number} deltaY
+   * @param {number} veloX
+   * @param {number} veloY
+   * @return {!Promise}
+   * @private
+   */
+  onZoomRelease_(centerClientX, centerClientY, deltaX, deltaY, veloX, veloY) {
+    let promise;
+    if (veloX == 0 && veloY == 0) {
+      promise = Promise.resolve();
+    } else {
+      promise = continueMotion(this.image_,
+          deltaX, deltaY, veloX, veloY,
+          (x, y) => {
+            this.onZoomInc_(centerClientX, centerClientY, x, y);
+            return true;
+          }).thenAlways();
+    }
+
+    const relayout = this.scale_ > this.startScale_;
+    return promise.then(() => {
+      return this.release_();
+    }).then(() => {
+      if (relayout) {
+        this.updateSrc_();
+      }
+    });
+  }
+
+  /**
+   * Sets or animates pan/zoom parameters.
+   * @param {number} newScale
+   * @param {number} newPosX
+   * @param {number} newPosY
+   * @param {boolean} animate
+   * @return {!Promise|undefined}
+   * @private
+   */
+  set_(newScale, newPosX, newPosY, animate) {
+    const ds = newScale - this.scale_;
+    const dx = newPosX - this.posX_;
+    const dy = newPosY - this.posY_;
+    const dist = Math.sqrt(dx * dx + dy * dy);
+
+    let dur = 0;
+    if (animate) {
+      const maxDur = 250;
+      dur = Math.min(maxDur, Math.max(
+          maxDur * dist * 0.01,      // Moving component.
+          maxDur * Math.abs(ds)));   // Zooming component.
+    }
+
+    let promise;
+    if (dur > 16 && animate) {
+      /** @const {!TransitionDef<number>} */
+      const scaleFunc = tr.numeric(this.scale_, newScale);
+      /** @const {!TransitionDef<number>} */
+      const xFunc = tr.numeric(this.posX_, newPosX);
+      /** @const {!TransitionDef<number>} */
+      const yFunc = tr.numeric(this.posY_, newPosY);
+      promise = Animation.animate(this.image_, time => {
+        this.scale_ = scaleFunc(time);
+        this.posX_ = xFunc(time);
+        this.posY_ = yFunc(time);
+        this.updatePanZoom_();
+      }, dur, PAN_ZOOM_CURVE_).thenAlways(() => {
+        this.scale_ = newScale;
+        this.posX_ = newPosX;
+        this.posY_ = newPosY;
+        this.updatePanZoom_();
+      });
+    } else {
+      this.scale_ = newScale;
+      this.posX_ = newPosX;
+      this.posY_ = newPosY;
+      this.updatePanZoom_();
+      if (animate) {
+        promise = Promise.resolve();
+      } else {
+        promise = undefined;
+      }
+    }
+
+    return promise;
+  }
+
+  /**
+   * Sets or animates pan/zoom parameters after release of the gesture.
+   * @return {!Promise}
+   * @private
+   */
+  release_() {
+    const newScale = this.boundScale_(this.scale_, false);
+    if (newScale != this.scale_) {
+      this.updatePanZoomBounds_(newScale);
+    }
+    const newPosX = this.boundX_(this.posX_ / this.scale_ * newScale, false);
+    const newPosY = this.boundY_(this.posY_ / this.scale_ * newScale, false);
+    return this.set_(newScale, newPosX, newPosY, true).then(() => {
+      this.startScale_ = this.scale_;
+      this.startX_ = this.posX_;
+      this.startY_ = this.posY_;
+    });
+  }
+}
+
